fullscreen looks like working
